pyproject.toml
[project]
name = "fastapi-bigger-apps"
version = "0.1.0"
description = "A simple following along of fastapi's bigger apps tutorial"
authors = [{ name = "tomit4", email = "mosssap@gmail.com" }]
dependencies = ["fastapi>=0.112.1", "uvicorn>=0.30.6", "python-dotenv>=1.0.1"]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_bigger_apps"]

[project.scripts]
dev = "fastapi_bigger_apps.main:main"
